Reduce number of `sudo` invocations
Every invocation gets logged to syslog/journal. With a high check
frequency, this causes quite a bit of noise. There are four identical
`fail2ban-client status $JAIL` invocations here that can be
consolidated.

Use `mapfile` to read the status output into memory. We could just use
a tempfile, too, but we'd have to take care not to leave it behind
when something goes wrong.